add booking status
