+++ src/main/java/com/vn/controller/GeneralController.java
@@ -75,7 +75,7 @@ public class GeneralController {
         return "redirect:/home";
     }
 
-    @GetMapping("/signin")
+    @GetMapping("/login")
     public String signIn() {
         return "home/home_guest";
     }
